+++ spec/integration_spec.rb
@@ -242,16 +242,16 @@ describe_cli 'jazzy' do
     end
 
     describe 'Creates docs for Swift project from symbolgraph files' do
-      build_path = Dir.getwd + 'tmp/.build'
+      build_path = Dir.getwd + '/tmp/.build'
       package_path =
         ROOT + 'spec/integration_specs/misc_jazzy_symgraph_features/before'
-      "swift build --package-path #{package_path} --build-path #{build_path} " \
-        "-Xswiftc -emit-symbol-graph -Xswiftc -emit-symbol-graph-dir -Xswiftc #{build_path}"
+      `swift build --package-path #{package_path} --build-path #{build_path} \
+        -Xswiftc -emit-symbol-graph -Xswiftc -emit-symbol-graph-dir -Xswiftc #{build_path}`
       module_path = `swift build --build-path #{build_path} --show-bin-path`
 
       behaves_like cli_spec 'misc_jazzy_symgraph_features',
                             '--swift-build-tool symbolgraph ' \
-                              "--symbolgraph_directory #{build_path} "
+                              "--symbolgraph-directory #{build_path} "
     end
   end if !spec_subset || spec_subset == 'swift'